+++ pcs_test/tier0/common/reports/test_messages.py
@@ -5560,3 +5560,19 @@ class AgentSelfValidationInvalidData(NameBuildTest):
             f"Invalid validation data from agent: {reason}",
             reports.AgentSelfValidationInvalidData(reason),
         )
+
+
+class BoothAuthfileNotUsed(NameBuildTest):
+    def test_message(self):
+        self.assert_message_from_report(
+            "Booth authfile is not enabled",
+            reports.BoothAuthfileNotUsed("instance name"),
+        )
+
+
+class BoothUnsupportedOptionEnableAuthfile(NameBuildTest):
+    def test_message(self):
+        self.assert_message_from_report(
+            "Unsupported option 'enable-authfile' is set in booth configuration",
+            reports.BoothUnsupportedOptionEnableAuthfile("instance name"),
+        )